// packages/@effects/src/views/setting/gnzjgl/ggcxtc/types.ts
export interface GgcxtcVO {
  id: number
  cNum: string
  cTitle: string
  nCbtype: string
  cCbclass: string
  cCondition: string
  cWidth: string
  cHeight: string
  nRows: number
  cSql: string
  cTablecols: string
  cCrenumb: string
  cCrename: string
  dDate: string
  cOrder: string
  nSfcommon: string
}